major rework of site navigation based on angular routing
Previously, each main section (address, sponsors...etc.) were a page.
This forced the reloading of the entire site, instanciations of
controllers...etc.

Rework entirely so that only index is loaded as a page, and then
everything else is a template managed by angular.

Improve texts internationalisation on client side.

Moreover, take the oportunity to clean.